refactor: split host editor general tab
